app/requirements.txt
cryptography
Flask==2.2.2
Flask-MySQL==1.5.1
PyMySQL==0.10.0
setuptools-rust
pyannote.audio
speechbrain
pydub
scipy
soundfile

